EVTV Official GEVCU Release
This is the release of GEVCU software currently shipping with EVTV
GEVCU Hardware.  This is the only release we can support or answer
questions on at EVTV.  There are several branches and variants of GEVCU
and we delighted to see it adapting and growing for a variety of
purposes.  However, we can only write documentation and provide support
for one.  This is it.

If you have questions about other variants, go to
www.github.com/collin80/GEVCU and try to determine the authors of the
particular branch or flavor you enjoy.